+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -6,7 +6,7 @@ The Button component allows us to initiate actions, make state or page changes.
 
 #### 🎉 New features
 
-- Added `maxWidth` property to the component. ([#322](https://github.com/TiendaNube/nimbus-design-system/pull/322) by [@joacotornello](https://github.com/joacotornello))
+- Added `fullWidth` property to the component. ([#322](https://github.com/TiendaNube/nimbus-design-system/pull/322) by [@joacotornello](https://github.com/joacotornello))
 
 ## 2025-07-10 `2.6.1`